requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart